Complete model updates with stronger properties.
The stronger temporal / reference properties for the ARV and WM are
complete and prove fully.

In order to make the above happen for the WM, I had to build a new
version of the theory of sequences that does not require the use of
uninterpreted functions. This new theory is less interesting from a
theoretic perspective but should not meaningfully impact the
expressivity of the UxAS model. Importantly, however, by eliminating the
use of uninterpreted functions, IC3 is able to run on the WM and the
full uxas model - this is critical to proving the strengthened temporal
property for the WM. Moreover, the proof of the waypoint-on-route
property is now possible without the need for any lemmas that would
reference internal state in their antecedents.

The full UxAS model is now able to reprove the properties of the WM and
ARV. Many new environment properties have been added, mostly as lemmas
for attempts to prove the critical compositional property that
mission-command messages reference, transively, previously received
unique automation response messages. This property, however, still
cannot be proved and is commented out in this commit.

All non-commented-out properties prove in both monolithic and
contract-based mode. All meaningful antecedents are observed using
witnesses.