Refactor uxas
This commit refactors uxas following advice from @daniel-larraz to use
observers to group ancillary properties in uxas. This makes choosing
properties to prove much easier than commenting/uncommenting and speeds
up analysis.

Curiously, all of the proofs in the Task fail if the contract is
"Activated" (currently, it is disabled).

In part because of this, a new service is added, empty_service, that
provably does nothing (always sends no_message). This allows the mode to
retain the infrastructrue for the task service without actually using
the task service.