Add a proof-of-concept model of the RA.
Following the presentation at the CRDInAL TIM, Dr. Jen Davis from
Collins asked about how the route aggregator was actually implemented in
lustre. Her question got me to thinking about how this might be done,
using the facilities in the sequences theory.

The answer is ... it works ... kind of. The details are in the comments
to `pocs/route-aggregator.lus`.

The property of interest proves in about 7 seconds using IC3 - but, as
the comments emphasize, the model is *highly* overconstrained because of
the way in which the `subsequence` function is defined.