Improve the model with recent kind2 features.
This commit uses the `check ... provided` synatx to improve the handling
of sequent-style checks. With this new syntax, the reachability check
for the antecent is generated automatically by kind2, eliminating the
need for a separate check. This makes the model more readable and
reduces the potential for error.

This commit also refactors and improves `task.lus` to use the
frame-condition syntax, which is much better suited for describing
what's going on in the task. The result is more readable and more likely
to be correct.